feat(ui): refactor EditableSectionList
