+++ app/shared/components/accordion-blocks/editable-section-list/EditableSectionList.tsx
@@ -2,9 +2,9 @@ import { DragEndEvent } from '@dnd-kit/core';
 import { Box, Typography } from '@mui/material';
 import { JSONContent } from '@tiptap/react';
 
-import { SortableItemWrapper } from '../../sortable-item-wrapper/SortableItemWrapper';
 import { SortableList } from '../../sortable-list/SortableList';
 import { styles } from './EditableSectionList.styles';
+import { EditableSectionListItem } from './EditableSectionListItem';
 import ConfigurableList from '~/components/configurable-list/ConfigurableList';
 import { CustomTextField } from '~/components/design-system/text-field/TextField';
 import { ConfigurableListItem } from '~/types/accordionBlocks';
@@ -35,39 +35,17 @@ export const EditableSectionList = <ItemType extends SectionListItem>({
   sectionLabel,
   onDragEnd
 }: EditableSectionListProps<ItemType>) => {
-  const renderItem = ({ item }: { item: ItemType }) => {
-    const fields = (
-      <Box display="flex" flexDirection="column" gap="16px" width="100%">
-        <CustomTextField
-          fieldType="formatting"
-          label="Заголовок пункту"
-          value={item.title}
-          onChange={(value) => onChangeItem(item.id as string, 'title', value)}
-        />
-        <CustomTextField
-          fieldType="formatting"
-          label="Текст пункту"
-          value={item.description}
-          onChange={(value) => onChangeItem(item.id as string, 'description', value)}
-        />
-      </Box>
-    );
-
-    if (onDragEnd) {
-      return (
-        <SortableItemWrapper id={item.id as string} key={item.id} gripHandle gripPosition="top">
-          {fields}
-        </SortableItemWrapper>
-      );
-    }
-
-    return fields;
-  };
 
   const listContent = (
     <ConfigurableList
       items={items}
-      renderItem={renderItem}
+      renderItem={({ item }: { item: ItemType }) => (
+        <EditableSectionListItem
+          item={item}
+          onChangeItem={onChangeItem}
+          onDragEnd={onDragEnd}
+        />
+      )}
       addBtnLabel="Додати пункт"
       onChange={({ id, field, value }) => onChangeItem(id as string, field, value)}
       onCreate={onCreateItem}
